cluster-unification: split TimelyConfig out of mz-compute-client
