+++ src/compute-client/Cargo.toml
@@ -18,6 +18,7 @@ futures = "0.3.25"
 http = "0.2.8"
 itertools = "0.10.5"
 mz-build-info = { path = "../build-info" }
+mz-cluster-client = { path = "../cluster-client" }
 mz-expr = { path = "../expr" }
 mz-orchestrator = { path = "../orchestrator" }
 mz-ore = { path = "../ore", features = ["tracing_"] }
