+++ src/compute-client/build.rs
@@ -106,6 +106,7 @@ fn main() {
         .extern_path(".mz_repr.relation_and_scalar", "::mz_repr")
         .extern_path(".mz_repr.row", "::mz_repr")
         .extern_path(".mz_repr.url", "::mz_repr::url")
+        .extern_path(".mz_cluster_client", "::mz_cluster_client")
         .extern_path(".mz_storage_client", "::mz_storage_client")
         .compile_with_config(
             config,
