+++ src/compute-client/src/protocol/command.rs
@@ -10,7 +10,6 @@
 //! Compute protocol commands.
 
 use std::collections::BTreeSet;
-use std::num::NonZeroI64;
 
 use proptest::prelude::{any, Arbitrary};
 use proptest::strategy::{BoxedStrategy, Strategy, Union};
@@ -19,6 +18,7 @@ use serde::{Deserialize, Serialize};
 use timely::progress::frontier::Antichain;
 use uuid::Uuid;
 
+use mz_cluster_client::client::{ClusterStartupEpoch, TimelyConfig};
 use mz_expr::RowSetFinishing;
 use mz_ore::tracing::OpenTelemetryContext;
 use mz_persist_client::cfg::PersistParameters;
@@ -60,7 +60,7 @@ pub enum ComputeCommand<T = mz_repr::Timestamp> {
     /// they are in.
     CreateTimely {
         config: TimelyConfig,
-        epoch: ComputeStartupEpoch,
+        epoch: ClusterStartupEpoch,
     },
 
     /// `CreateInstance` must be sent after `CreateTimely` to complete the [Creation Stage] of the
@@ -343,124 +343,6 @@ impl Arbitrary for ComputeCommand<mz_repr::Timestamp> {
                     })
                     .boxed(),
             ])
-    }
-}
-
-/// A value generated by environmentd and passed to the clusterd processes
-/// to help them disambiguate different `CreateTimely` commands.
-///
-/// The semantics of this value are not important, except that they
-/// must be totally ordered, and any value (for a given replica) must
-/// be greater than any that were generated before (for that replica).
-/// This is the reason for having two
-/// components (one from the stash that increases on every environmentd restart,
-/// another in-memory and local to the current incarnation of environmentd)
-#[derive(PartialEq, Eq, Debug, Copy, Clone, Serialize, Deserialize)]
-pub struct ComputeStartupEpoch {
-    envd: NonZeroI64,
-    replica: u64,
-}
-
-impl RustType<ProtoComputeStartupEpoch> for ComputeStartupEpoch {
-    fn into_proto(&self) -> ProtoComputeStartupEpoch {
-        let Self { envd, replica } = self;
-        ProtoComputeStartupEpoch {
-            envd: envd.get(),
-            replica: *replica,
-        }
-    }
-
-    fn from_proto(proto: ProtoComputeStartupEpoch) -> Result<Self, TryFromProtoError> {
-        let ProtoComputeStartupEpoch { envd, replica } = proto;
-        Ok(Self {
-            envd: envd.try_into().unwrap(),
-            replica,
-        })
-    }
-}
-
-impl ComputeStartupEpoch {
-    pub fn new(envd: NonZeroI64, replica: u64) -> Self {
-        Self { envd, replica }
-    }
-
-    /// Serialize for transfer over the network
-    pub fn to_bytes(&self) -> [u8; 16] {
-        let mut ret = [0; 16];
-        let mut p = &mut ret[..];
-        use std::io::Write;
-        p.write_all(&self.envd.get().to_be_bytes()[..]).unwrap();
-        p.write_all(&self.replica.to_be_bytes()[..]).unwrap();
-        ret
-    }
-
-    /// Inverse of `to_bytes`
-    pub fn from_bytes(bytes: [u8; 16]) -> Self {
-        let envd = i64::from_be_bytes((&bytes[0..8]).try_into().unwrap());
-        let replica = u64::from_be_bytes((&bytes[8..16]).try_into().unwrap());
-        Self {
-            envd: envd.try_into().unwrap(),
-            replica,
-        }
-    }
-}
-
-impl std::fmt::Display for ComputeStartupEpoch {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        let Self { envd, replica } = self;
-        write!(f, "({envd}, {replica})")
-    }
-}
-
-impl PartialOrd for ComputeStartupEpoch {
-    fn partial_cmp(&self, other: &Self) -> Option<std::cmp::Ordering> {
-        Some(self.cmp(other))
-    }
-}
-
-impl Ord for ComputeStartupEpoch {
-    fn cmp(&self, other: &Self) -> std::cmp::Ordering {
-        let Self { envd, replica } = self;
-        let Self {
-            envd: other_envd,
-            replica: other_replica,
-        } = other;
-        (envd, replica).cmp(&(other_envd, other_replica))
-    }
-}
-
-/// Configuration of the cluster we will spin up
-#[derive(Arbitrary, Clone, Debug, Default, PartialEq, Serialize, Deserialize)]
-pub struct TimelyConfig {
-    /// Number of per-process worker threads
-    pub workers: usize,
-    /// Identity of this process
-    pub process: usize,
-    /// Addresses of all processes
-    pub addresses: Vec<String>,
-    /// The amount of effort to be spent on arrangement compaction during idle times.
-    ///
-    /// See [`differential_dataflow::Config::idle_merge_effort`].
-    pub idle_arrangement_merge_effort: u32,
-}
-
-impl RustType<ProtoTimelyConfig> for TimelyConfig {
-    fn into_proto(&self) -> ProtoTimelyConfig {
-        ProtoTimelyConfig {
-            workers: self.workers.into_proto(),
-            addresses: self.addresses.into_proto(),
-            process: self.process.into_proto(),
-            idle_arrangement_merge_effort: self.idle_arrangement_merge_effort,
-        }
-    }
-
-    fn from_proto(proto: ProtoTimelyConfig) -> Result<Self, TryFromProtoError> {
-        Ok(Self {
-            process: proto.process.into_rust()?,
-            workers: proto.workers.into_rust()?,
-            addresses: proto.addresses.into_rust()?,
-            idle_arrangement_merge_effort: proto.idle_arrangement_merge_effort,
-        })
     }
 }
