+++ src/compute-client/src/service.rs
@@ -24,12 +24,13 @@ use timely::PartialOrder;
 use tonic::{Request, Status, Streaming};
 use uuid::Uuid;
 
+use mz_cluster_client::client::TimelyConfig;
 use mz_repr::{Diff, GlobalId, Row};
 use mz_service::client::{GenericClient, Partitionable, PartitionedState};
 use mz_service::grpc::{GrpcClient, GrpcServer, ProtoServiceTypes, ResponseStream};
 
 use crate::metrics::ReplicaMetrics;
-use crate::protocol::command::{ComputeCommand, ProtoComputeCommand, TimelyConfig};
+use crate::protocol::command::{ComputeCommand, ProtoComputeCommand};
 use crate::protocol::response::{
     ComputeResponse, PeekResponse, ProtoComputeResponse, SubscribeBatch, SubscribeResponse,
 };
